Revert "Refactor/update share info"
