+++ components/organisms/result/MyResult.tsx
@@ -6,7 +6,9 @@ import Average from 'components/atoms/emissions/Average'
 import Cloud from 'components/atoms/emissions/Cloud'
 import DatasourceFooter from 'components/DatasourceFooter'
 import QuestionResultGraph from 'components/molecules/result/ResultGraph'
+import ShareSNS from 'components/molecules/result/ShareSNS/ShareSNS'
 import { useEmissionResult } from 'hooks/emission'
+import { useProfile } from 'hooks/profile'
 
 type Props = {
   category: Questions.QuestionCategory
@@ -15,6 +17,7 @@ type Props = {
 export const MyResult: FC<Props> = ({ category }) => {
   const router = useRouter()
   const result = useEmissionResult(category)
+  const { profile } = useProfile()
 
   const sortedResult = useMemo(() => {
     const r = result[category]
@@ -29,6 +32,22 @@ export const MyResult: FC<Props> = ({ category }) => {
     const r = result[category]
     return r ? Math.round(r.find((m) => m.key === 'total')?.value || 0) : 0
   }, [result])
+
+  const additional_hashtag = process.env.NEXT_PUBLIC_TWITTER_SHARE_TAG
+    ? `,${process.env.NEXT_PUBLIC_TWITTER_SHARE_TAG}`
+    : ''
+
+  const twitterShareLink = useMemo(() => {
+    return `https://twitter.com/share?url=${process.env.NEXT_PUBLIC_CLIENT_URL}/category/${category}/result/${profile?.shareId}&text=わたしのカーボンフットプリント量&hashtags=じぶんごとプラネット${additional_hashtag}`
+  }, [profile, category])
+
+  const facebookShareLink = useMemo(() => {
+    return `https://www.facebook.com/sharer/sharer.php?u=${process.env.NEXT_PUBLIC_CLIENT_URL}/category/${category}/result/${profile?.shareId}`
+  }, [profile, category])
+
+  const lineShareLink = useMemo(() => {
+    return `https://line.me/R/msg/text/?${process.env.NEXT_PUBLIC_CLIENT_URL}/category/${category}/result/${profile?.shareId}`
+  }, [profile, category])
 
   return (
     <>
@@ -54,6 +73,12 @@ export const MyResult: FC<Props> = ({ category }) => {
       )}
 
       <QuestionResultGraph category={category} sortedResult={sortedResult} />
+
+      <ShareSNS
+        facebook={facebookShareLink}
+        line={lineShareLink}
+        twitter={twitterShareLink}
+      />
 
       <Text fontWeight="bold" textAlign="center" mt={10} mb={5}>
         カーボンフットプリント量を減らすために
